chore: some tidying
